Replace code with the official version 6.1.32
[DESCRIPTION]

Needed because there are no tracking information in
the original SVN repository, when a new release was
created.

[FILE]
*

+++ src/VBox/Devices/EFI/Firmware/DynamicTablesPkg/Include/StandardNameSpaceObjects.h
@@ -79,7 +79,7 @@ typedef struct CmAStdObjAcpiTableInfo {
   UINT32                         AcpiTableSignature;
 
   /// The ACPI table revision
-  UINT8                          AcpiTableRevision;
+  UINT32                         AcpiTableRevision;
 
   /// The ACPI Table Generator ID
   ACPI_TABLE_GENERATOR_ID        TableGeneratorId;
