+++ src/VBox/Runtime/common/checksum/RTSha1Digest.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTSha1Digest.cpp 93301 2022-01-18 11:24:43Z vboxsync $ */
+/* $Id: RTSha1Digest.cpp $ */
 /** @file
  * IPRT - SHA1 digest creation
  *
@@ -7,7 +7,7 @@
  */
 
 /*
- * Copyright (C) 2009-2022 Oracle Corporation
+ * Copyright (C) 2009-2020 Oracle Corporation
  *
  * This file is part of VirtualBox Open Source Edition (OSE), as
  * available from http://www.virtualbox.org. This file is free software;
@@ -57,7 +57,7 @@ RTR3DECL(int) RTSha1Digest(void* pvBuf, size_t cbBuf, char **ppszDigest, PFNRTPR
     RTSha1Init(&Ctx);
 
     /* Buffer size for progress callback */
-    double rdMulti = 100.0 / (cbBuf ? (double)cbBuf : 1.0);
+    double rdMulti = 100.0 / (cbBuf ? cbBuf : 1);
 
     /* Working buffer */
     char *pvTmp = (char*)pvBuf;
@@ -74,7 +74,7 @@ RTR3DECL(int) RTSha1Digest(void* pvBuf, size_t cbBuf, char **ppszDigest, PFNRTPR
         /* Call the progress callback if one is defined */
         if (pfnProgressCallback)
         {
-            rc = pfnProgressCallback((unsigned)((double)cbReadTotal * rdMulti), pvUser);
+            rc = pfnProgressCallback((unsigned)(cbReadTotal * rdMulti), pvUser);
             if (RT_FAILURE(rc))
                 break; /* canceled */
         }
@@ -119,7 +119,7 @@ RTR3DECL(int) RTSha1DigestFromFile(const char *pszFile, char **ppszDigest, PFNRT
         return rc;
 
     /* Fetch the file size. Only needed if there is a progress callback. */
-    double rdMulti = 0.0;
+    double rdMulti = 0;
     if (pfnProgressCallback)
     {
         uint64_t cbFile;
@@ -129,7 +129,7 @@ RTR3DECL(int) RTSha1DigestFromFile(const char *pszFile, char **ppszDigest, PFNRT
             RTFileClose(hFile);
             return rc;
         }
-        rdMulti = 100.0 / (cbFile ? (double)cbFile : 1.0);
+        rdMulti = 100.0 / (cbFile ? cbFile : 1);
     }
 
     /* Allocate a reasonably large buffer, fall back on a tiny one. */
@@ -160,7 +160,7 @@ RTR3DECL(int) RTSha1DigestFromFile(const char *pszFile, char **ppszDigest, PFNRT
         /* Call the progress callback if one is defined */
         if (pfnProgressCallback)
         {
-            rc = pfnProgressCallback((unsigned)((double)cbReadTotal * rdMulti), pvUser);
+            rc = pfnProgressCallback((unsigned)(cbReadTotal * rdMulti), pvUser);
             if (RT_FAILURE(rc))
                 break; /* canceled */
         }
